iio:adc:ad5592r: Fixed buffer bug
The iio_push_to_buffers expects a buffer with nr_active_scan samples
to be pushed simultaneously to the iio_buffer. Thus a buffer needs
to be set up which receives the samples in the for loop after which it
will be loaded in to the iio_buffer using the iio_push_to_buffers
function.